+++ src/main/java/lotto/domain/User.java
@@ -1,9 +1,10 @@
-package lotto;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import camp.nextstep.edu.missionutils.Randoms;
+import lotto.utils.Constants;
 
 public class User {
 	private Integer money;
